Rename modelfitting package to fitter
