+++ brian2modelfitting/fitter.py
